requirements.txt
Django==1.10
Pygments==2.1.3
appnope==0.1.0
decorator==4.0.10
djangorestframework==3.4.3
httpie==0.9.6
ipython==5.0.0
ipython-genutils==0.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.5
psycopg2==2.6.2
ptyprocess==0.5.1
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
